change structure project
