+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screen/constants/colors.dart';
+import 'package:flutter_screen/ui/constants/colors.dart';
 import 'package:flutter_screen/ui/profile/profile.dart';
 import 'package:flutter_screen/ui/settings/settingsScreen.dart';
 import 'package:flutter_screen/ui/topsection/topsection.dart';
@@ -46,7 +46,7 @@ class _MyHomePageState extends State<MyHomePage> {
           headerSliverBuilder: (context, innerBoxInScrolled) =>
               [sliverAppBar()],
           body: const TabBarView(
-            children: [profileScreen(), settingsScreen()],
+            children: [ProfileScreen(), settingsScreen()],
           )),
     );
   }
