+++ lib/ui/profile/profile.dart
@@ -1,22 +1,22 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screen/ui/profile/carousel.dart';
-import 'package:flutter_screen/ui/profile/chips.dart';
-import 'package:flutter_screen/ui/profile/listTiles.dart';
-import 'package:flutter_screen/ui/profile/subTextSection.dart';
-import 'package:flutter_screen/ui/profile/tarifsText.dart';
-import 'package:flutter_screen/ui/profile/textSection.dart';
+import 'package:flutter_screen/ui/profile/sections/carousel.dart';
+import 'package:flutter_screen/ui/profile/sections/chips.dart';
+import 'package:flutter_screen/ui/profile/sections/filtersTextSection.dart';
+import 'package:flutter_screen/ui/profile/sections/listTiles.dart';
+import 'package:flutter_screen/ui/profile/sections/subTextSection.dart';
+import 'package:flutter_screen/ui/profile/sections/tarifsTextSection.dart';
 
-class profileScreen extends StatelessWidget {
-  const profileScreen({super.key});
+class ProfileScreen extends StatelessWidget {
+  const ProfileScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
     return ListView(padding: EdgeInsets.zero, children: const [
-      subscriptions(),
-      carousel(),
-      tarifksText(),
-      listTiles(),
-      filtersText(),
+      SubsText(),
+      Carousel(),
+      TarifsAndLimitsText(),
+      ListTiles(),
+      FiltersText(),
       Chips()
     ]);
   }
