+++ lib/ui/topsection/topsection.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screen/constants/colors.dart';
-import 'package:flutter_screen/constants/paddings.dart';
-import 'package:flutter_screen/constants/strings.dart';
-import 'package:flutter_screen/constants/sizes.dart';
+import 'package:flutter_screen/ui/constants/colors.dart';
+import 'package:flutter_screen/ui/constants/paddings.dart';
+import 'package:flutter_screen/ui/constants/strings.dart';
+import 'package:flutter_screen/ui/constants/sizes.dart';
 
 SliverAppBar sliverAppBar() {
   return SliverAppBar(
